ReactionCommons

This package is part of the ReactionEngine

Documentation for ReactionCommons.

ReactionCommons.Arrhenius

ReactionCommons.AuxiliaryData

ReactionCommons.GasphaseMechanism

ReactionCommons.GasphaseReaction

ReactionCommons.GasphaseRxnStoichiometry

ReactionCommons.MechanismDefinition

ReactionCommons.Order

ReactionCommons.SiteInfo

ReactionCommons.Sri

ReactionCommons.Troe

ReactionCommons.UserDefinedState

ReactionCommons.get_dependencies

ReactionCommons.parse_rxn_string

ReactionCommons.species_rxn_map

ReactionCommons.Arrhenius — Type

Arrhenius reaction rate parameters

ReactionCommons.AuxiliaryData — Type

Auxiliary reaction data for gasphase kinetics

ReactionCommons.GasphaseMechanism — Type

Object defining the gasphase reaction mechanism

ReactionCommons.GasphaseReaction — Type

Object defining an elementary gasphase reaction

ReactionCommons.GasphaseRxnStoichiometry — Type

Stoichiometry object for gasphase reactions for an individual reaction

ReactionCommons.MechanismDefinition — Type

An anstract type for surface reaction mechanism parameter definitions

ReactionCommons.Order — Type

Object for defining order of a reaction using FORD or RORD keyword for gasphase reactions

ReactionCommons.SiteInfo — Type

composite type for the definition of site definition of a surface reaction mechanism

ReactionCommons.Sri — Type

Parameters for Sri reaction parameters

ReactionCommons.Troe — Type

Parameters for troe reaction parameters

ReactionCommons.UserDefinedState — Type

User state for use in user defined reaction rate calculations

ReactionCommons.get_dependencies — Method

Function to determine the coverage dependency, order dependency and MWC for a reaction

Usage

get_dependencies(cov_dep_rxns,order_dep_rxns,mwc_rxns,rxn_id)

covdeprxns::Dict{Int64=>Dict{Int64=>Float64}}

orderdeprxns::Dict{Int64=>Dict{Int64=>Float64}}

mwc_rxns::Array{Int64}

rxn_id::Int64

The function returns a tuple of cov,order and mwc

ReactionCommons.parse_rxn_string — Method

Separate the reactants and products from the stoichiometric equation

Usage:

parase_rxn_string(rxn_str)

rxn_str::String: String representing the reaction equation

ReactionCommons.species_rxn_map — Method

Create array of structures of SpeciesRxnMap. The SpeciesRxnMap is a composit structure of species_id in participating reactions and the stoichiometric coefficients

Usage

species_rxn_map(gas_species,sm)

gas_species::Array{String,1} : list of gas phase species

sm::SurfaceMechanism : SurfaceMechanism composite type